Add files via upload
John and I worked together today. We created demo html  files for each of the categories we need to style-- tables, forms, buttons, etc. And then we broke the CSS file into segments with comments.  We created classes for most of the basic text styles, like underline and italic. John is going to work on buttons now and I'm working on forms. 
We added some typography styles, as a starting point, and also put in some media queries (with temporary colored backgrounds, just to make sure the queries are working). The breakpoints can be easily changed.